requirements.txt
Django==1.8.2
django-ckeditor==4.4.8
django-grappelli==2.6.5
